Added more info about what the firmware is
I've been seeing lots of questions about the closed source status of the
firmware. Some of it is people conflating the responsibilities of our
firmware/system management processor vs the SMPs of other similar
products. So I decided to write a small explination about what our FW is
doing and that we will be open-sourcing what we can in the future.