Renaming .tar.gz to .fwbundle to avoid confusion
Folks have a natural predelection for seeing a .tar.gz (or a .zip, etc)
and immediately opening it fully.  This has led to some slight confusion
since we aren't expecting folks to actually open this up.  Moving this
to a non-standard suffix but nothing else is really changing.  This move
precedes the additional tt-flash option to add an additional option to
also take a fwbundle file

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>